spp: expand configuration fragments included by relative path
Some configuration fragments are included via a relative path, i.e.
include ../../feature2/feature.cfg

This breaks the relocation code, as the name of the fragment
will never match any of the path prefixs.

We solve that by using the kconfig name as located by searching the
includes. Those are always absolute and will match the prefix, allowing
us to figure out the relocation directory.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>