kgit-s2q: add --prev flag to retry a patch
The way that kgit-s2q was previously used would always skip a
failed patch if you reran the series.

That is not always desireable if you fix a patch, and want to
have it retried on the next run.

Adding a --prev flag to the script which will roll back the
fence post files to a state where the last patch will be
retried versus skipped.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>