tools: allow fixups to conditionally change options
Sometimes we have options that should either be =y or =m depending
on the use case.

Rather than force the options into multiple parallel and very similar
fragments (which is error prone and labour intensive), we can allow
an annotation in the configuration fragment that allows the default
value to be overriden by a variable.

The variable in question comes from a -D<var>=<value> on the scc
or spp command line. If the define evaluates to a non-zero value
a fixup is created that will modify the value when scc finishes
gathering meta-data.

An example of a notation would be:

  CONFIG_INET_TUNNEL=y # OVERRIDE:$MODULE_OR_Y

CONFIG_INET_TUNNEL will be =y when standard tools (like merge-config)
are used. But when scc/spp see this notation, they check the value
of the variable MODULE_OR_Y, if that variable evalutes to a non-zero
value, a fixup is created that will change CONFIG_INET_TUNNEL to
that value before the kernel is configured.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>