kgit-s2q: fix commit SHA reproducibility with raw diffs
In case of git-am failure, kgit-s2q tries to read the date from the
patch. Unfortunately, if the patch does not have a date (e.g. raw diff),
no date is set and git uses the current system time.

This patch makes kgit-s2q use SOURCE_DATE_EPOCH as the last resort or emit
a warning if it's not available. Additionally, GIT_COMMITTER_DATE is set
to GIT_AUTHOR_DATE if requested with --commit-sha option (same behavior as
with git-am).

Signed-off-by: Hubert Wiśniewski <hubert.wisniewski.25632@gmail.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>